requirements.txt
alembic==1.0.11
appdirs==1.4.3
atomicwrites==1.3.0
attrs==19.1.0
bcrypt==3.1.7
black==19.3b0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
dataclasses==0.6
entrypoints==0.3
fastapi==0.35.0
flake8==3.7.8
h11==0.8.1
httptools==0.0.13
idna==2.8
importlib-metadata==0.19
Mako==1.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.2.0
packaging==19.1
passlib==1.7.1
pluggy==0.12.0
py==1.8.0
pycodestyle==2.5.0
pydantic==0.30
pyflakes==2.1.1
PyJWT==1.7.1
pyparsing==2.4.2
pytest==5.0.1
pytest-cov==2.7.1
python-dateutil==2.8.0
python-editor==1.0.4
python-multipart==0.0.5
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.7
starlette==0.12.7
toml==0.10.0
urllib3==1.25.3
uvicorn==0.8.6
uvloop==0.12.2
wcwidth==0.1.7
websockets==7.0
zipp==0.5.2